entities and repo changed

+++ employee-management-hrms/employee-management-hrms/src/main/java/com/employee/management/hrms/repository/EmployeeRepository.java
@@ -11,7 +11,7 @@ public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
 
     Optional<Employee> findByEmployeeCode(String employeeCode);
 
-    List<Employee> findByDepartment_DepartmentId(Integer departmentId);
+    List<Employee> findByDepartment_DepartmentId(Long departmentId);
 
-    List<Employee> findByManager_EmployeeId(Integer managerId);
+    List<Employee> findByManager_EmployeeId(Long managerId);
 }
